src/requirements.txt
altgraph==0.17.4
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
et_xmlfile==2.0.0
fonttools==4.58.4
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
MouseInfo==0.1.3
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywin32==310
pywin32-ctypes==0.2.3
ratelim==0.1.6
reportlab==4.4.1
requests==2.32.4
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
tkintermapview==1.29
urllib3==2.4.0
